DonantesKevin/fundacion-mascotas/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosqlite==0.19.0
httpx==0.25.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
aiosqlite==0.19.0
